Fix up CodeFactor issues
